Add: Hello World Type Node
Issue #6! Start! Here we add a "helloWorld" type node in which I experiment with some basic output of directory (Wanting to make sure the info we printed in python is useable in cpp! <3 So, after that in should be a simple matter of getting the client and server going tmrw! Signing off for now- n