Add: filepath call sucess!
Continues issue #6  - This update has the server return some of the client info as a check- it will the be used as a param in the xml parsing! Overall, many successful builds with the cmake now! (Note- using issue#6 database which will be included in the next commit for posterity)